% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{byars_upper}
\alias{byars_upper}
\title{byars_upper}
\usage{
byars_upper(x, confidence = 0.95)
}
\arguments{
\item{x}{the observed numbers of events; numeric vector; no default}
}
\value{
Returns upper confidence limits for observed numbers of events using Byar's method [1]
}
\description{
Calculates the upper confidence limits for observed numbers of events using Byar's method [1].
}
\section{Notes}{
 \code{\link{byars_lower}} and \code{byars_upper} together return symmetric confidence
 intervals around counts, therefore
 for a specified confidence level, \eqn{\alpha}, the probability that, by chance, the upper limit returned
 will be below the true underlying value, is \eqn{\alpha}/2.
}

\references{
[1] Breslow NE, Day NE. Statistical methods in cancer research,
 volume II: The design and analysis of cohort studies. Lyon: International
 Agency for Research on Cancer, World Health Organisation; 1987.

-------------------------------------------------------------------------------------------------
}
