% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DSR_data}
\alias{DSR_data}
\title{SII test datasets - DSR}
\format{A data table}
\usage{
data(DSR_data)
}
\description{
A data table of dummy Directly Standardised Rates by deprivation quintiles
}
\examples{
DSR_data

}
\keyword{datasets}
