% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esp2013}
\alias{esp2013}
\title{European Standard Population 2013}
\format{A numeric vector with 19 elements}
\usage{
esp2013
}
\description{
A numeric vector containing the nineteen 5-year age band populations making up the 2013 European Standard Population
ordered from age 0-4, 5-9, 10-14 ... to ...  85-89, 90+:
}
\details{
5000 5500 5500 5500 6000 6000 6500 7000 7000 7000 7000 6500 6000 5500 5000 4000 2500 1500 1000
}
\keyword{datasets}
