% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FindXValues}
\alias{FindXValues}
\title{FindXValues}
\usage{
FindXValues(xvals, no_quantiles)
}
\arguments{
\item{xvals}{field name in input dataset that contains the quantile
populations; unquoted string; no default}

\item{no_quantiles}{number of quantiles supplied in dataset for SII;
integer; no default}
}
\description{
Calculates mid-points of cumulative population for each quantile
}
