% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LE_data}
\alias{LE_data}
\title{SII test datasets - Life Expectancy}
\format{A data table}
\usage{
data(LE_data)
}
\description{
A data table of life expectancy data by area and deprivation decile
}
\examples{
LE_data

}
\keyword{datasets}
