% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funnel_prop_sig.R
\name{phe_fun_prop_sig}
\alias{phe_fun_prop_sig}
\title{Calculate confidence intervals/control limits and levels of significance}
\usage{
phe_fun_prop_sig(data, x, denom, multiplier = 100)
}
\arguments{
\item{data}{a data.frame containing the data to calculate control limits for,}

\item{x}{field name from data containing the observed numbers of cases in the sample meeting the required condition
(the numerator for the CLs); unquoted string; no default}

\item{denom}{field name from data containing the population(s) in the sample (the denominator for the CLs);
unquoted string; no default}

\item{multiplier}{the multiplier used to express the final values (eg 100 = percentage); numeric; default 1}
}
\value{
returns the original data.frame with the significance level appended
}
\description{
Calculates control limits adopting consistent method as per the PHE Fingertips Technical Guidance : https://fingertips.phe.org.uk/profile/guidance
}
\examples{
library(dplyr)
df<-data.frame(Area= c("A","B","C","D"),
               numerator = c(10232,12321,15123,13213),
               denominator = c(15232,16123,17932,18475))
df\%>\%
    phe_fun_prop_sig(numerator, denominator)


}
\seealso{
Other PHEindicatormethods package functions: 
\code{\link{phe_dsr}()},
\code{\link{phe_fun_prop_plot}()},
\code{\link{phe_isr}()},
\code{\link{phe_life_expectancy}()},
\code{\link{phe_mean}()},
\code{\link{phe_proportion}()},
\code{\link{phe_quantile}()},
\code{\link{phe_rate}()},
\code{\link{phe_sii}()},
\code{\link{phe_smr}()}
}
\concept{PHEindicatormethods package functions}
