% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funnels.R
\name{phe_funnel_convert_points}
\alias{phe_funnel_convert_points}
\title{Derive rate and annual population values for charting based. Process removes
rates where the rate type is dsr and the number of observed events are below
10.}
\usage{
phe_funnel_convert_points(
  data,
  numerator,
  denominator,
  rate,
  rate_type,
  years_of_data,
  multiplier = NULL
)
}
\arguments{
\item{data}{a data.frame containing the data to calculate control limits for;
unquoted string; no default}

\item{numerator}{field name from data containing the observed numbers of cases in the
sample meeting the required condition (the numerator for the CLs); unquoted
string; no default}

\item{denominator}{field name from data containing the population(s) in the
sample (the denominator for the CLs); unquoted string; no default}

\item{rate}{field name from data containing the rate data when creating
funnels for a Directly Standardised Rate; unquoted string; no default}

\item{rate_type}{string; one of "dsr" or "crude"; when statistic is rate this
argument is required}

\item{years_of_data}{numeric; number of years the data represents; this is
required for statistic = "rate"}

\item{multiplier}{the multiplier used to express the final values (eg 100 =
percentage); numeric; default 100}
}
\value{
returns the same table as provided with two additional fields. First
  will have the same name as the rate field, with the suffix "_chart", the
  second will be called denominator_derived
}
\description{
Derive rate and annual population values for charting based. Process removes
rates where the rate type is dsr and the number of observed events are below
10.
}
\author{
Sebastian Fox, \email{sebastian.fox@phe.gov.uk}
}
