% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funnels.R
\name{phe_funnel_significance}
\alias{phe_funnel_significance}
\title{Identifies level of outlier based on funnel plot methodology}
\usage{
phe_funnel_significance(
  data,
  x,
  denominator,
  multiplier = 100,
  statistic = "proportion"
)
}
\arguments{
\item{data}{a data.frame containing the data to calculate control limits for,}

\item{x}{field name from data containing the observed numbers of cases in the
sample meeting the required condition (the numerator for the CLs); unquoted
string; no default}

\item{denominator}{field name from data containing the population(s) in the
sample (the denominator for the CLs); unquoted string; no default}

\item{multiplier}{the multiplier used to express the final values (eg 100 =
percentage); numeric; default 100}

\item{statistic}{string; type of statistic to inform funnel calculations.
Currently only accepts "proportion"}
}
\value{
returns the original data.frame with the significance level appended
}
\description{
Determines whether records are outliers, and the level they are outliers,
among a dataset of numerators and denominators. This follows the same
methodology as that published on the PHE Fingertips Technical Guidance page:
https://fingertips.phe.org.uk/profile/guidance
}
\examples{
library(dplyr)
df <- data.frame(
  Area = c("A", "B", "C", "D"),
  numerator = c(10232, 12321, 15123, 13213),
  denominator = c(15232, 16123, 17932, 18475)
)
df \%>\%
  phe_funnel_significance(numerator, denominator)

}
\seealso{
Other PHEindicatormethods package functions: 
\code{\link{phe_dsr}()},
\code{\link{phe_funnels}()},
\code{\link{phe_isr}()},
\code{\link{phe_life_expectancy}()},
\code{\link{phe_mean}()},
\code{\link{phe_proportion}()},
\code{\link{phe_quantile}()},
\code{\link{phe_rate}()},
\code{\link{phe_sii}()},
\code{\link{phe_smr}()}
}
\author{
Matthew Francis, \email{Matthew.Francis@phe.gov.uk}
}
\concept{PHEindicatormethods package functions}
