% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funnel_prop_sig.R
\name{phe_funnels}
\alias{phe_funnels}
\title{Calculate confidence intervals/control limits and levels of significance}
\usage{
phe_funnels(data, x, denominator)
}
\arguments{
\item{data}{a data.frame containing the data to calculate control limits for;
unquoted string; no default}

\item{x}{field name from data containing the observed numbers of cases in the sample meeting the
required condition (the numerator for the CLs); unquoted string; no default}

\item{denominator}{field name from data containing the population(s) in the sample
(the denominator for the CLs); unquoted string; no default}
}
\value{
returns the original data.frame with the following appended:
        lower 0.025 limit, upper 0.025 limit, lower 0.001 limit, upper 0.001 limit and
        baseline average
}
\description{
Calculates control limits adopting consistent method as per the PHE Fingertips Technical Guidance : https://fingertips.phe.org.uk/profile/guidance
}
\seealso{
Other PHEindicatormethods package functions: 
\code{\link{phe_dsr}()},
\code{\link{phe_funnel_significance}()},
\code{\link{phe_isr}()},
\code{\link{phe_life_expectancy}()},
\code{\link{phe_mean}()},
\code{\link{phe_proportion}()},
\code{\link{phe_quantile}()},
\code{\link{phe_rate}()},
\code{\link{phe_sii}()},
\code{\link{phe_smr}()}
}
\concept{PHEindicatormethods package functions}
