% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMR.R
\name{phe_smr}
\alias{phe_smr}
\title{phe_smr}
\usage{
phe_smr(data, x, n, x_ref, n_ref, type = "standard", confidence = 0.95,
  ref = 1)
}
\arguments{
\item{data}{data.frame containing the data to be standarised, pre-grouped if multiple SMRs required; unquoted string; no default}

\item{x}{field name from data containing the observed number of events for each standardisation category (eg ageband) within each grouping set (eg area);
unquoted string; no default}

\item{n}{field name from data containing the populations for each standardisation category (eg ageband) within each grouping set (eg area);
unquoted string; no default}

\item{x_ref}{the observed number of events in the reference population for each standardisation category
(eg age band) within each grouping set (eg area). If reference populations are held within data,
the column must be referenced as a vector and not repeated eg df$x_ref[1:19]; numeric vector; no default}

\item{n_ref}{the reference population for each standardisation category
(eg age band) within each grouping set (eg area).  If reference populations are held within data,
the column must be referenced as a vector and not repeated eg df$n_ref[1:19]; numeric vector; no default}

\item{type}{type of output; can be "value", "lower", "upper", "standard" (for all 3 previous fields) or "full"; quoted string; default combined}

\item{confidence}{the required level of confidence expressed as a number between 0.9 and 1
or 90 and 100; numeric; default 0.95}

\item{ref}{the standardised reference ratio, numeric, default = 1}
}
\value{
When type = "full", returns a data frame of observed, expected, value, lowercl, uppercl, confidence, statistic and method
        for each grouping set
}
\description{
Calculates a standard mortality ratio (or indirectly standardised ratio) with confidence limits using Byars or Exact CI method.
}
\examples{
NEED TO ADD EXAMPLES

}
\seealso{
Other PHEstatmethods package functions: \code{\link{byars_lower}},
  \code{\link{byars_upper}}, \code{\link{phe_dsr}},
  \code{\link{phe_isr}}, \code{\link{phe_mean}},
  \code{\link{phe_proportion}}, \code{\link{phe_rate}},
  \code{\link{wilson_lower}}, \code{\link{wilson_upper}},
  \code{\link{zscore}}
}
