% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHEindicatormethods.R
\docType{package}
\name{PHEindicatormethods}
\alias{PHEindicatormethods}
\title{PHEindicatormethods: A package for performing standard statistics for public
health indicators}
\description{
PHEindicatormethods: A package for performing standard statistics for public
health indicators
}
