% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prevalence_data}
\alias{prevalence_data}
\title{SII test datasets - Prevalence}
\format{
A data table
}
\usage{
data(prevalence_data)
}
\description{
A data table of example prevalence data by area and deprivation decile
}
\examples{
prevalence_data

}
\keyword{datasets}
