% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funnels.R
\name{sigma_adjustment}
\alias{sigma_adjustment}
\title{Calculate the funnel point value for a specific population based on a
population average value}
\usage{
sigma_adjustment(p, population, average_proportion, side, multiplier = 100)
}
\arguments{
\item{p}{numeric (between 0 and 1); probability to calculate funnel plot
point (will normally be either 0.975 or 0.999)}

\item{population}{numeric; the population for the area}

\item{average_proportion}{numeric; the average proportion for all the areas
included in the funnel plot (the sum of the numerators divided by the sum
of the denominators)}

\item{side}{string; "low" or "high" to determine which funnel to calculate}

\item{multiplier}{numeric; the multiplier used to express the final values
(eg 100 = percentage); default 100}
}
\value{
returns a value equivalent to the specified funnel for the input
  population
}
\description{
Returns a value equivalent to the higher/lower funnel plot point based on the
input population and probability
}
\author{
Sebastian Fox, \email{sebastian.fox@phe.gov.uk}
}
