% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilson.R
\name{wilson_lower}
\alias{wilson_lower}
\title{Wilson}
\usage{
wilson_lower(x, n, confidence = 0.95)
}
\arguments{
\item{x}{the observed number of cases in the sample meeting the required condition; numeric vector; no default}

\item{n}{the number of cases in the sample; numeric vector; no default}

\item{confidence}{the required level of confidence expressed as a number between 0.9 and 1
or 90 and 100; numeric; default 0.95}
}
\value{
Returns a lower confidence limit for an observed number of events usign Wilson's method
}
\description{
Calculates lower confidence limit for a proportion using Wilson's method.
}
\examples{
wilson_lower(65,100)
wilson_lower(65,100,99.8)

}
\seealso{
\code{\link{phe_proportion}} for proportions,
         \code{\link{phe_rate}} for rates,
         \code{\link{phe_mean}} for means,
         \code{\link{phe_dsr}} for directly standardised rates,
         \code{\link{phe_isr}} for indirectly standardised ratios/rates and standardised mortality ratios

Other phe statistical functions: \code{\link{byars_lower}},
  \code{\link{byars_upper}}, \code{\link{phe_dsr}},
  \code{\link{phe_isr}}, \code{\link{phe_mean}},
  \code{\link{phe_proportion}}, \code{\link{phe_rate}},
  \code{\link{wilson_upper}}
}
