% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zscore.R
\name{zscore}
\alias{zscore}
\title{zscore}
\usage{
zscore(confidence = 0.95)
}
\arguments{
\item{confidence}{the required level of confidence expressed as a number between 0.9 and 1
or 90 and 100; numeric; default 0.95}
}
\value{
Returns a z score for the required percentile value from the Standard Normal distirbution
}
\description{
Calculates z scores (percentile values) from the Standard Normal distribution
}
\examples{
zscore(0.95)
zscore(99.8)

}
\seealso{
Other PHEstatmethods package functions: \code{\link{byars_lower}},
  \code{\link{byars_upper}}, \code{\link{phe_dsr}},
  \code{\link{phe_isr}}, \code{\link{phe_mean}},
  \code{\link{phe_proportion}}, \code{\link{phe_rate}},
  \code{\link{phe_smr}}, \code{\link{wilson_lower}},
  \code{\link{wilson_upper}}
}
