% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byars.R
\name{byars_lower}
\alias{byars_lower}
\title{Calculates lower confidence limit for an observed number of events using Byar's method.}
\usage{
byars_lower(x, conf.level = 0.95)
}
\arguments{
\item{x}{the observed number of events; numeric; no default}

\item{conf.level}{the level of confidence required expressed as a number between 0.9 and 1
or 90 and 100; numeric; default 0.95}
}
\value{
Returns a lower confidence limit
}
\description{
Calculates lower confidence limit for an observed number of events using Byar's method.
}
\examples{
byars_lower(65)
byars_lower(65,99.8)
}
