% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byars.R
\name{byars_upper}
\alias{byars_upper}
\title{byars_upper}
\usage{
byars_upper(x, confidence = 0.95)
}
\arguments{
\item{x}{the observed number of events; numeric vector; no default}

\item{confidence}{the level of confidence required expressed as a number between 0.9 and 1
or 90 and 100; numeric; default 0.95}
}
\value{
Returns an upper confidence limit for an observed number of events using Byar's method
}
\description{
Calculates the upper confidence limit for an observed number of events using Byar's method.
}
\examples{
byars_upper(65)
byars_upper(65,99.8)

}
\seealso{
Other PHEstatmethods package functions: \code{\link{byars_lower}},
  \code{\link{phe_dsr}}, \code{\link{phe_mean}},
  \code{\link{phe_proportion}}, \code{\link{phe_rate}},
  \code{\link{wilson_lower}}, \code{\link{wilson_upper}}
}
