% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSR.R
\name{phe_dsr}
\alias{phe_dsr}
\title{DSR}
\usage{
phe_dsr(data, x, n, stdpop, type = "standard", confidence = 0.95,
  multiplier = 1e+05)
}
\arguments{
\item{data}{data.frame containing the data to be standarised, pre-grouped if multiple DSRs required; unquoted string; no default}

\item{x}{field name from data containing the observed number of events for each standardisation category (eg ageband) within each grouping set (eg area);
unquoted string; no default}

\item{n}{field name from data containing the populations for each standardisation category (eg ageband) within each grouping set (eg area);
unquoted string; no default}

\item{stdpop}{the standard populations for each standardisation category (eg age band); unquoted numeric vector; no default}

\item{type}{type of output; can be "value", "lower", "upper", "standard" (for all 3 previous fields to be added to your output) or "full"; string; default combined}

\item{confidence}{the required level of confidence expressed as a number between 0.9 and 1
or 90 and 100; numeric; default 0.95}

\item{multiplier}{the multiplier used to express the final values (eg 100,000 = rate per 100,000,
100 = percentage); numeric; default 100,000}
}
\value{
When type = "full" returns a data.frame of numerator, denominator, directly standardised rate,
        lower and upper confidence limits and method
}
\description{
Calculates a directly standardised rate with confidence limits using Dobson method.
}
\examples{
library(dplyr)
df <- data.frame(indicatorid = rep(c(1234, 5678, 91011, 121314), each = 19 * 2 * 5),
                 year = rep(2006:2010, each = 19 * 2),
                 sex = rep(rep(c("Male", "Female"), each = 19), 5),
                 obs = sample(200, 19 * 2 * 5 * 4, replace = TRUE),
                 pop = sample(10000:20000, 19 * 2 * 5 * 4, replace = TRUE))
df \%>\%
    group_by(indicatorid, year, sex) \%>\%
    phe_dsr(obs, pop, esp2013)

## OR

df \%>\%
    group_by(indicatorid, year, sex) \%>\%
    phe_dsr(obs, pop, esp2013, "full")

}
\seealso{
\code{\link{phe_proportion}} for proportions,
         \code{\link{phe_rate}} for rates,
         \code{\link{phe_mean}} for means,
         \code{\link{phe_dsr}} for directly standardised rates,
         \code{\link{phe_isr}} for indirectly standardised ratios/rates and standardised mortality ratios

Other phe statistical functions: \code{\link{byars_lower}},
  \code{\link{byars_upper}}, \code{\link{phe_isr}},
  \code{\link{phe_mean}}, \code{\link{phe_proportion}},
  \code{\link{phe_rate}}, \code{\link{wilson_lower}},
  \code{\link{wilson_upper}}
}
