% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSR.R, R/SMR.R
\name{phe_dsr}
\alias{phe_dsr}
\alias{phe_dsr}
\title{Calculates a directly standardised rate with confidence limits using Dobson method.}
\usage{
phe_dsr(x, n, stdpop = esp2013, conf.level = 0.95, multiplier = 1e+05)

phe_dsr(x, n, stdpop = esp2013, conf.level = 0.95, multiplier = 1e+05)
}
\arguments{
\item{x}{the observed number of events for each standardisation category (eg ageband) within each group (eg area);
numeric vector; no default}

\item{n}{the populations for each standardisation category (eg ageband) within each group (eg area);
numeric vector; no default}

\item{stdpop}{the standard populations for each standardisation category
 (eg age band); numeric vector; default is the European Standard Population 2013
divided into 19 five-year agebands: 0-4, 5-9, 10-14, .....90+}

\item{conf.level}{the required level of confidence expressed as a number between 0.9 and 1
or 90 and 100; numeric; default 0.95}

\item{multiplier}{the multiplier used to express the final values (eg 100,000 = rate per 100,000,
100 = percentage); numeric; default 100,000}

\item{groupref}{the}

\item{x}{the observed number of events for each standardisation category (eg age band);
numeric vector; no default}

\item{n}{the population for each standardisation category (eg age band);
numeric vector; no default}

\item{ref_x}{the observed number(s) of events in the reference population for each standardisation
category
 (eg age band); numeric vector; default is the European Standard Population 2013
with 19 five-year age band categories}

\item{conf.level}{the required level of confidence expressed as a number between 0.9 and 1
or 90 and 100; numeric; default 0.95}

\item{multiplier}{the multiplier used to express the final values (eg 100,000 = rate per 100,000,
100 = percentage); numeric; default 100,000}
}
\value{
Returns a data frame of method, numerator, denominator, directly standardised rate
        and confidence interval limits

Returns a data frame of method, numerator, denominator, directly standardised rate
        and confidence interval limits
}
\description{
Calculates a directly standardised rate with confidence limits using Dobson method.

Calculates an indirectly standardised rate with confidence limits using Byar's method.
}
\examples{
phe_dsr(c(27,45,55,100,125,300,295,270,275,450,455,459,345,300,270,265,100,90,35),
        c(84935,80367,72122,79259,99806,87362,81579,71103,70001,
          69007,63203,52638,46087,40887,32604,28399,21625,13021,7355),conf.level = 0.998)
phe_dsr(c(27,45,55,100,125,300,295,270,275,450,455,459,345,300,270,265,100,90,35),
        c(84935,80367,72122,79259,99806,87362,81579,71103,70001,
          69007,63203,52638,46087,40887,32604,28399,21625,13021,7355),conf.level = 0.998)
}
