% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSR.R
\name{phe_dsr}
\alias{phe_dsr}
\title{Calculates a directly standardised rate with confidence limits using Dobson method.}
\usage{
phe_dsr(x, n, stdpop, groupref = "No Grouping", conf.level = 0.95,
  multiplier = 100000)
}
\arguments{
\item{x}{the observed number of events for each standardisation category (eg ageband) within each grouping set (eg area);
numeric vector; no default}

\item{n}{the populations for each standardisation category (eg ageband) within each grouping set (eg area);
numeric vector; no default}

\item{stdpop}{the standard populations for each standardisation category
 (eg age band); numeric vector; no default; the European Standard Population 2013
divided into 19 five-year agebands (0-4, 5-9, 10-14, .....90+) is available within the package}

\item{groupref}{the grouping sets (eg area codes or area names) if calculating multiple DSRs at once,
character vector, default = No Grouping}

\item{conf.level}{the required level of confidence expressed as a number between 0.9 and 1
or 90 and 100; numeric; default 0.95}

\item{multiplier}{the multiplier used to express the final values (eg 100,000 = rate per 100,000,
100 = percentage); numeric; default 100,000}
}
\value{
Returns a data frame of numerator, denominator, directly standardised rate,
        lower and upper confidence limits and method
}
\description{
Calculates a directly standardised rate with confidence limits using Dobson method.
}
\examples{
phe_dsr(c(27,45,55,100,125,300,295,270,275,450,455,459,345,300,270,265,100,90,35),
        c(84935,80367,72122,79259,99806,87362,81579,71103,70001,
          69007,63203,52638,46087,40887,32604,28399,21625,13021,7355), esp2013, conf.level = 0.998)
}
\seealso{
\code{\link{phe_proportion}} for proportions,
         \code{\link{phe_rate}} for rates,
         \code{\link{phe_mean}} for means,
         \code{\link{phe_dsr}} for directly standardised rates,
         \code{\link{phe_isr}} for indirectly standardised ratios/rates and standardised mortality ratios

Other phe statistical functions: \code{\link{byars_lower}},
  \code{\link{byars_upper}}, \code{\link{phe_isr}},
  \code{\link{phe_mean}}, \code{\link{phe_proportion}},
  \code{\link{phe_rate}}
}
