% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funnels.R
\name{phe_funnel_significance}
\alias{phe_funnel_significance}
\title{Identifies level of outlier based on funnel plot methodology}
\usage{
phe_funnel_significance(
  data,
  numerator,
  denominator,
  rate,
  statistic = "proportion",
  rate_type = NA,
  multiplier = NULL
)
}
\arguments{
\item{data}{a data.frame containing the data to calculate control limits for,}

\item{numerator}{field name from data containing the observed numbers of cases in the
sample meeting the required condition (the numerator for the CLs); unquoted
string; no default}

\item{denominator}{field name from data containing the population(s) in the
sample (the denominator for the CLs); unquoted string; no default}

\item{rate}{field name from data containing the rate data when creating
funnels for a Directly Standardised Rate; unquoted string; no default}

\item{statistic}{string; type of statistic to inform funnel calculations.
Acceptable values are "proportion", "ratio" or "rate"}

\item{rate_type}{string; one of "dsr" or "crude"; when statistic is rate this
argument is required}

\item{multiplier}{numeric; the multiplier that the rate is normalised
with (ie, per 100,000); only required when statistic = "rate}
}
\value{
returns the original data.frame with the significance level appended
}
\description{
Determines whether records are outliers, and the level they are outliers,
among a dataset of numerators and denominators. This follows the same
methodology as that published on the PHE Fingertips Technical Guidance page:
https://fingertips.phe.org.uk/profile/guidance
}
\examples{
library(dplyr)
df <- data.frame(
  Area = c("A", "B", "C", "D"),
  numerator = c(10232, 12321, 15123, 13213),
  denominator = c(15232, 16123, 17932, 18475)
)
df \%>\%
  phe_funnel_significance(numerator, denominator)

}
\seealso{
Other PHEindicatormethods package functions: 
\code{\link{phe_dsr}()},
\code{\link{phe_funnels}()},
\code{\link{phe_isr}()},
\code{\link{phe_life_expectancy}()},
\code{\link{phe_mean}()},
\code{\link{phe_proportion}()},
\code{\link{phe_quantile}()},
\code{\link{phe_rate}()},
\code{\link{phe_sii}()},
\code{\link{phe_smr}()}
}
\author{
Matthew Francis, \email{Matthew.Francis@phe.gov.uk}
}
\concept{PHEindicatormethods package functions}
