% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funnels.R
\name{phe_funnels}
\alias{phe_funnels}
\title{Calculate confidence intervals/control limits for funnel plots}
\usage{
phe_funnels(
  data,
  numerator,
  denominator,
  rate,
  type = "full",
  multiplier = 100,
  statistic = "proportion",
  ratio_type = NULL,
  rate_type = NULL,
  years_of_data = NULL
)
}
\arguments{
\item{data}{a data.frame containing the data to calculate control limits for;
unquoted string; no default}

\item{numerator}{field name from data containing the observed numbers of cases in the
sample meeting the required condition (the numerator for the CLs); unquoted
string; no default}

\item{denominator}{field name from data containing the population(s) in the
sample (the denominator for the CLs); unquoted string; no default}

\item{rate}{field name from data containing the rate data when creating
funnels for a Directly Standardised Rate; unquoted string; no default}

\item{type}{defines the data and metadata columns to be included in output;
"standard" (for all data) or "full" (for all data and metadata); quoted
string; default = "full"}

\item{multiplier}{the multiplier used to express the final values (eg 100 =
percentage); numeric; default 100}

\item{statistic}{string; type of statistic to inform funnel calculations.
Acceptable values are "proportion", "ratio" or "rate"}

\item{ratio_type}{if statistic is "ratio" the user can specify either "count"
or "isr" (indirectly standardised rate) here as a string}

\item{rate_type}{string; one of "dsr" or "crude"; when statistic is rate this
argument is required}

\item{years_of_data}{numeric; number of years the data represents; this is
required for statistic = "rate"}
}
\value{
returns the original data.frame with the following appended: lower
  0.025 limit, upper 0.025 limit, lower 0.001 limit, upper 0.001 limit and
  baseline average
}
\description{
Calculates control limits adopting a consistent method as per the PHE
Fingertips Technical Guidance: https://fingertips.phe.org.uk/profile/guidance
}
\examples{
library(dplyr)
set.seed(123)
df <- data.frame(obs = sample(200, 19 * 2 * 5 * 4, replace = TRUE),
                 pop = sample(10000:20000, 19 * 2 * 5 * 4, replace = TRUE))
df \%>\%
    phe_funnels(obs, pop)

}
\seealso{
Other PHEindicatormethods package functions: 
\code{\link{phe_dsr}()},
\code{\link{phe_funnel_significance}()},
\code{\link{phe_isr}()},
\code{\link{phe_life_expectancy}()},
\code{\link{phe_mean}()},
\code{\link{phe_proportion}()},
\code{\link{phe_quantile}()},
\code{\link{phe_rate}()},
\code{\link{phe_sii}()},
\code{\link{phe_smr}()}
}
\author{
Matthew Francis, \email{Matthew.Francis@phe.gov.uk}
}
\concept{PHEindicatormethods package functions}
