% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISR.R
\name{phe_isr}
\alias{phe_isr}
\title{phe_isr}
\usage{
phe_isr(data, x, n, x_ref, n_ref, type = "standard", confidence = 0.95,
  multiplier = 1e+05)
}
\arguments{
\item{data}{data.frame containing the data to be standarised, pre-grouped if multiple ISRs required; unquoted string; no default}

\item{x}{field name from data containing the observed number of events for each standardisation category (eg ageband) within each grouping set (eg area);
unquoted string; no default}

\item{n}{field name from data containing the populations for each standardisation category (eg ageband) within each grouping set (eg area);
unquoted string; no default}

\item{x_ref}{the observed number of events in the reference population for each standardisation category
(eg age band) within each grouping set (eg area). If reference populations are held within data,
the column must be referenced as a vector and not repeated eg df$x_ref[1:19]; numeric vector; no default}

\item{n_ref}{the reference population for each standardisation category
(eg age band) within each grouping set (eg area).  If reference populations are held within data,
the column must be referenced as a vector and not repeated eg df$n_ref[1:19]; numeric vector; no default}

\item{type}{type of output; can be "value", "lower", "upper", "standard" (for all 3 previous fields) or "full"; quoted string; default combined}

\item{confidence}{the required level of confidence expressed as a number between 0.9 and 1
or 90 and 100; numeric; default 0.95}

\item{multiplier}{the multiplier used to express the final values (eg 100,000 = rate per 100,000); numeric; default 100,000}
}
\value{
When type = "full", returns a data frame of observed, expected, ref_rate, value, lowercl, uppercl, confidence, statistic and method
        for each grouping set
}
\description{
Calculates an indirectly standardised rate with confidence limits using Byars or Exact CI method.
}
\examples{
NEED TO ADD EXAMPLES

}
\seealso{
Other PHEstatmethods package functions: \code{\link{byars_lower}},
  \code{\link{byars_upper}}, \code{\link{phe_dsr}},
  \code{\link{phe_mean}}, \code{\link{phe_proportion}},
  \code{\link{phe_rate}}, \code{\link{phe_smr}},
  \code{\link{wilson_lower}}, \code{\link{wilson_upper}}
}
