% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISR.R
\name{phe_isr}
\alias{phe_isr}
\title{Calculates an indirectly standardised ratio or rate with confidence limits using Byars or Exact CI method.}
\usage{
phe_isr(x, n, x_ref, n_ref, groupref = "No Grouping", conf.level = 0.95,
  ratio = TRUE, ref_0 = TRUE, multiplier = 1e+05)
}
\arguments{
\item{x}{the observed number of events for each standardisation category (eg ageband) within each grouping set (eg area);
numeric vector; no default}

\item{n}{the populations for each standardisation category (eg ageband) within each grouping set (eg area);
numeric vector; no default}

\item{x_ref}{the observed number of events in the reference population for each standardisation category
(eg age band) within each grouping set (eg area); numeric vector; no default}

\item{n_ref}{the reference population for each standardisation category
(eg age band) within each grouping set (eg area); numeric vector; no default}

\item{groupref}{the grouping sets (eg area codes or area names) if calculating multiple ISRs at once;
character vector; default = No Grouping}

\item{conf.level}{the required level of confidence expressed as a number between 0.9 and 1
or 90 and 100; numeric; default 0.95}

\item{ratio}{whether the output should be expressed as a ratio (TRUE) or a rate (FALSE); logical; default = TRUE}

\item{ref_0}{when ratio = TRUE - whether the ratios should be compared to a reference ratio of 0 or 100; logical, default=TRUE}

\item{multiplier}{when ratio = FALSE - the multiplier used to express the final ISRs (eg 100,000 = rate per 100,000,
100 = percentage); numeric; default 100,000}
}
\value{
When ratio = TRUE   - returns a data frame of row label, observed count, expected count,
                            indirectly standardised rate, lower and upper confidence interval limits and method
        When ratio = FALSE  - returns a data frame of row label, observed count, expected count,
                            indirectly standardised rate, lower and upper confidence interval limits, reference rate, and method
}
\description{
Calculates an indirectly standardised ratio or rate with confidence limits using Byars or Exact CI method.
}
\examples{
phe_isr(c(27,45,55,100,125,300,295,270,275,450,455,459,345,300,270,265,100,90,35),
        c(84935,80367,72122,79259,99806,87362,81579,71103,70001,
          69007,63203,52638,46087,40887,32604,28399,21625,13021,7355),
        c(10303,2824,3225,3615,3641,3490,3789,3213,3031,2771,3089,3490,3595,4745,5514,7125,5694,6210,5757),
        c(50520,57173,60213,54659,44345,50128,62163,67423,62899,55463,60479,49974,
          44140,40888,37239,30819,18136,15325,13918))

}
