% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Means.R
\name{phe_mean}
\alias{phe_mean}
\title{Calculates a mean with confidence limits using Student's-t distribution method.}
\usage{
phe_mean(x, groupref = "No grouping", conf.level = 0.95)
}
\arguments{
\item{x}{the observed values in the sample(s)/population(s); numeric vector; no default}

\item{groupref}{the grouping sets (eg area codes or area names) if calculating multiple means at once,
character vector, default = No Grouping}

\item{conf.level}{the required level of confidence expressed as a number between 0.9 and 1
or 90 and 100; numeric; default 0.95}
}
\value{
Returns a data frame of row labels, sum of values, count of values, mean, standard deviation,
        lower and upper confidence limits and method
}
\description{
Calculates a mean with confidence limits using Student's-t distribution method.
}
\examples{
phe_mean(c(20,30,40), 0.95)

## Example of the grouping parameter
df <- data.frame(group = rep(letters[1:5], each = 5),
                 value = runif(25))
phe_mean(df$value, df$group)
}
\seealso{
\code{\link{phe_proportion}} for proportions,
         \code{\link{phe_rate}} for rates,
         \code{\link{phe_mean}} for means,
         \code{\link{phe_dsr}} for directly standardised rates,
         \code{\link{phe_isr}} for indirectly standardised ratios/rates and standardised mortality ratios

Other phe statistical functions: \code{\link{byars_lower}},
  \code{\link{byars_upper}}, \code{\link{phe_dsr}},
  \code{\link{phe_isr}}, \code{\link{phe_proportion}},
  \code{\link{phe_rate}}
}
