% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proportions.R
\name{phe_proportion}
\alias{phe_proportion}
\title{Calculates a proportion with confidence limits using Wilson method.}
\usage{
phe_proportion(x, n, conf.level = 0.95, percentage = FALSE)
}
\arguments{
\item{x}{a vector of numerators; numeric; no default}

\item{n}{a vector of denominators; numeric; no default}

\item{conf.level}{the level of confidence required expressed as a number between 0.9 and 1
or 90 and 100; numeric; default 0.95}

\item{percentage}{whether the output should be returned as a percentage; logical; default FALSE}
}
\value{
Returns a data frame of method, numerator, denominator, proportion and confidence limits
}
\description{
Calculates a proportion with confidence limits using Wilson method.
}
\examples{
phe_proportion(65,100)
phe_proportion(65,100,99.8,TRUE)

}
\seealso{
\code{\link{phe_proportion}} for proportions,
         \code{\link{phe_rate}} for rates,
         \code{\link{phe_dsr}} for directly standardised rates,
         \code{\link{phe_isr}} for indirectly standardised rates and standardised mortality ratios
}
