% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rates.R
\name{phe_rate}
\alias{phe_rate}
\title{Calculates a rate with confidence limits using Byar's method.}
\usage{
phe_rate(x, n, conf.level = 0.95, multiplier = 1e+05)
}
\arguments{
\item{x}{the numerator; numeric; no default}

\item{n}{the denominator; numeric; no default}

\item{conf.level}{the level of confidence required expressed as a number between 0.9 and 1
or 90 and 100; numeric; default 0.95}

\item{multiplier}{the multiplier used to express the final values (eg 100,000 = rate per 100,000,
100 = percentage); numeric; default 100,000}
}
\value{
Returns a data frame of method, numerator, denominator, rate and confidence limits
}
\description{
Calculates a rate with confidence limits using Byar's method.
}
\examples{
phe_rate(65,100)
phe_rate(65,100,99.8,100)
}
