% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rates.R
\name{phe_rate}
\alias{phe_rate}
\title{Rate}
\usage{
phe_rate(data, x, n, type = "standard", confidence = 0.95,
  multiplier = 1e+05)
}
\arguments{
\item{data}{the data.frame containing the data to calculate rates for; unquoted string; no default}

\item{x}{field name from data containing the rate numerators (eg observed number of events); unquoted string; no default}

\item{n}{field name from data containing the rate denominators (eg populations); unquoted string; no default}

\item{type}{type of output; can be "value", "lower", "upper", "standard" (for all 3 previous fields to be added to your output) or "full"; string; default combined}

\item{confidence}{the required level of confidence expressed as a number between 0.9 and 1
or 90 and 100; numeric; default 0.95}

\item{multiplier}{the multiplier used to express the final values (eg 100,000 = rate per 100,000,
100 = percentage); numeric; default 100,000}
}
\value{
When type=full, returns the original data.frame with the following columns appended:
        rate, lower confidence limit, upper confidence limit, confidence level, statistic and method
}
\description{
Calculates a rate with confidence limits using Byar's or Exact CI method.
}
\examples{
library(dplyr)
df <- data.frame(area = rep(c("Area1","Area2","Area3","Area4"), 2),
                 year = rep(2015:2016, each = 4),
                 obs = sample(100, 2 * 4, replace = TRUE),
                 pop = sample(100:200, 2 * 4, replace = TRUE))
phe_rate(df, obs, pop)

}
\seealso{
\code{\link{phe_proportion}} for proportions,
         \code{\link{phe_rate}} for rates,
         \code{\link{phe_mean}} for means,
         \code{\link{phe_dsr}} for directly standardised rates,
         \code{\link{phe_isr}} for indirectly standardised ratios/rates and standardised mortality ratios

Other phe statistical functions: \code{\link{byars_lower}},
  \code{\link{byars_upper}}, \code{\link{phe_dsr}},
  \code{\link{phe_isr}}, \code{\link{phe_mean}},
  \code{\link{phe_proportion}}, \code{\link{wilson_lower}},
  \code{\link{wilson_upper}}
}
