% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wilson_lower}
\alias{wilson_lower}
\title{wilson_lower}
\usage{
wilson_lower(x, n, confidence = 0.95)
}
\value{
Returns lower confidence limits for observed numbers of events using the Wilson Score method [1,2]
}
\description{
Calculates lower confidence limits for observed numbers of events using the Wilson Score method [1,2].
}
\section{Notes}{
 \code{wilson_lower} and \code{\link{wilson_upper}} together return symmetric confidence
 intervals, therefore for a specified confidence level, \eqn{\alpha}, the probability that, by chance, the lower limit returned
 will be above the true underlying value, is \eqn{\alpha}/2.#'
}

\references{
[1] Wilson EB. Probable inference, the law of succession, and statistical
 inference. J Am Stat Assoc; 1927; 22. Pg 209 to 212. \cr
[2] Newcombe RG, Altman DG. Proportions and their differences. In Altman
 DG et al. (eds). Statistics with confidence (2nd edn). London: BMJ Books;
 2000. Pg 46 to 48.

------------------------------------------------------------------------------------------------
}
