% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilson.R
\name{wilson_lower}
\alias{wilson_lower}
\title{wilson_lower}
\usage{
wilson_lower(x, n, confidence = 0.95)
}
\arguments{
\item{x}{the observed numbers of cases in the samples meeting the required condition; numeric vector; no default}

\item{n}{the numbers of cases in the samples; numeric vector; no default}

\item{confidence}{the required level of confidence expressed as a number between 0.9 and 1
or 90 and 100; numeric; default 0.95}
}
\value{
Returns lower confidence limits for observed numbers of events using the Wilson Score method [1,2]
}
\description{
Calculates lower confidence limits for observed numbers of events using the Wilson Score method [1,2].
}
\section{Notes}{
 \code{wilson_lower} and \code{\link{wilson_upper}} together return symmetric confidence
 intervals, therefore for a specified confidence level, \eqn{\alpha}, the probability that, by chance, the lower limit returned
 will be above the true underlying value, is \eqn{\alpha}/2.#'
}

\examples{
wilson_lower(65,100)
wilson_lower(65,100,99.8)

}
\references{
[1] Wilson EB. Probable inference, the law of succession, and statistical
 inference. J Am Stat Assoc; 1927; 22. Pg 209 to 212. \cr
[2] Newcombe RG, Altman DG. Proportions and their differences. In Altman
 DG et al. (eds). Statistics with confidence (2nd edn). London: BMJ Books;
 2000. Pg 46 to 48.
}
\seealso{
Other PHEindicatormethods package functions: \code{\link{byars_lower}},
  \code{\link{byars_upper}}, \code{\link{phe_dsr}},
  \code{\link{phe_isr}}, \code{\link{phe_mean}},
  \code{\link{phe_proportion}}, \code{\link{phe_rate}},
  \code{\link{phe_sii}}, \code{\link{phe_smr}},
  \code{\link{wilson_upper}}
}
\concept{PHEindicatormethods package functions}
