% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilson.R
\name{wilson_upper}
\alias{wilson_upper}
\title{wilson_upper}
\usage{
wilson_upper(x, n, confidence = 0.95)
}
\arguments{
\item{x}{the observed numbers of cases in the samples meeting the required condition; numeric vector; no default}

\item{n}{the numbers of cases in the samples; numeric vector; no default}

\item{confidence}{the required level of confidence expressed as a number between 0.9 and 1
or 90 and 100; numeric; default 0.95}
}
\value{
Returns upper confidence limits for observed numbers of events using the Wilson Score method
}
\description{
Calculates upper confidence limits for observed numbers of events using the Wilson Score method.
}
\examples{
wilson_upper(65,100)
wilson_upper(65,100,99.8)

}
\seealso{
Other PHEindicatormethods package functions: \code{\link{byars_lower}},
  \code{\link{byars_upper}}, \code{\link{phe_dsr}},
  \code{\link{phe_isr}}, \code{\link{phe_mean}},
  \code{\link{phe_proportion}}, \code{\link{phe_rate}},
  \code{\link{phe_smr}}, \code{\link{wilson_lower}}
}
